+++ streamdeck/src/plugin.ts
@@ -2,7 +2,7 @@ import streamDeck from "@elgato/streamdeck";
 
 import { HomeAction, MoveAction, PresetAction, ProfileAction, StatusAction, ZoomAction } from "./actions/camera";
 import { TrackingDialAction } from "./actions/dial";
-import { AutoZoomAction, LockAction, MotionSyncAction, TrackingAction } from "./actions/toggles";
+import { AnchorAction, AutoZoomAction, LockAction, MotionSyncAction, TrackingAction } from "./actions/toggles";
 import { initRaster } from "./render/raster";
 import { trackmind, type Connection } from "./trackmind";
 
@@ -12,6 +12,7 @@ streamDeck.actions.registerAction(new TrackingAction());
 streamDeck.actions.registerAction(new LockAction());
 streamDeck.actions.registerAction(new AutoZoomAction());
 streamDeck.actions.registerAction(new MotionSyncAction());
+streamDeck.actions.registerAction(new AnchorAction());
 streamDeck.actions.registerAction(new PresetAction());
 streamDeck.actions.registerAction(new HomeAction());
 streamDeck.actions.registerAction(new ProfileAction());